refactor(satellite): move common assertion
