+++ src/libs/satellite/src/assets/cdn/assert.rs
@@ -1,9 +1,5 @@
-use crate::assets::cdn::constants::{CDN_JUNO_PATH, CDN_JUNO_RELEASES_COLLECTION_KEY};
+use crate::assets::cdn::constants::CDN_JUNO_RELEASES_COLLECTION_KEY;
 use candid::Principal;
-use junobuild_cdn::storage::assert_releases_description;
-use junobuild_cdn::storage::errors::{
-    JUNO_CDN_STORAGE_ERROR_INVALID_COLLECTION, JUNO_CDN_STORAGE_ERROR_INVALID_RELEASES_PATH,
-};
 use junobuild_collections::assert::stores::{
     assert_create_permission, assert_create_permission_with, assert_permission,
     assert_permission_with,
@@ -12,49 +8,7 @@ use junobuild_collections::constants::assets::COLLECTION_ASSET_KEY;
 use junobuild_collections::types::core::CollectionKey;
 use junobuild_collections::types::rules::Permission;
 use junobuild_shared::controllers::{controller_can_write, is_controller};
-use junobuild_shared::regex::build_regex;
 use junobuild_shared::types::state::Controllers;
-use junobuild_storage::types::state::FullPath;
-
-// TODO: storage module uses cdn because of this reference. Refactor modules.
-
-pub fn assert_cdn_asset_keys(
-    full_path: &FullPath,
-    description: &Option<String>,
-    collection: &CollectionKey,
-) -> Result<(), String> {
-    match collection.as_str() {
-        CDN_JUNO_RELEASES_COLLECTION_KEY => {
-            assert_releases_keys(full_path)?;
-            assert_releases_description(description)?;
-
-            Ok(())
-        }
-        _ => {
-            if full_path.starts_with(CDN_JUNO_PATH) {
-                return Err(format!(
-                    "{} ({} - {})",
-                    JUNO_CDN_STORAGE_ERROR_INVALID_COLLECTION, full_path, collection
-                ));
-            }
-
-            Ok(())
-        }
-    }
-}
-
-fn assert_releases_keys(full_path: &FullPath) -> Result<(), String> {
-    let full_path_re = build_regex(r"^/_juno/releases/satellite[^/]*\.wasm\.gz$")?;
-
-    if !full_path_re.is_match(full_path) {
-        return Err(format!(
-            "{} ({})",
-            JUNO_CDN_STORAGE_ERROR_INVALID_RELEASES_PATH, full_path
-        ));
-    }
-
-    Ok(())
-}
 
 pub fn assert_cdn_write_on_dapp_collection(caller: Principal, controllers: &Controllers) -> bool {
     // When using proposals, we allow any controllers to upload - submit - an asset to be served from #dapp collection
