+++ src/libs/satellite/src/assets/mod.rs
@@ -1,2 +1,3 @@
+mod assert;
 pub mod cdn;
 pub mod storage;
