+++ src/libs/satellite/src/assets/storage/strategy_impls.rs
@@ -1,4 +1,4 @@
-use crate::assets::cdn::assert::assert_cdn_asset_keys;
+use crate::assets::assert::assert_cdn_asset_keys;
 use crate::assets::storage::assert::assert_storage_list_permission;
 use crate::assets::storage::state::{
     delete_asset, get_asset, get_config, get_domains, get_rule, insert_asset, insert_asset_encoding,
